rework shmem interface to be "symmetric"
The notion of "VILLASnode" / "external program" is removed from the
interface. Now there's only a distiction regarding which process opened
the object first, and this is handled transparently by the functions.

This enables some new uses for the interface, for example connecting
two instances of VILLASnode or two instances of DPsim directly.