Changed benchmarking
* node-infiniband-benchmark.sh now regenerates the config file every
loop. Furthermore, it does not save the source log anymore but instead
routes the generated signal to a file node.
* evaluate_logs.ipynb is now able to evaluate the signal output.
Furthermore the analysis of signal times and missed signals is enhanced
in this commit.

The plot already looks pretty good, but the alignment of text in
matplotlib is terrible, so there is still some work to do.