Fixed bug which improved latency about 33%.
An ugly workaround was removed in 3c33696265bcedf58efc207da716356e71682b97 and 159cefb14d9f951ae79ec8265c6bb0395009b19c. Besides making the code a little cleaner, it was expected that this would improve latency. However, it did not.

Cause for this was a bug that disabled inline messages for all messages. This is fixed in this commit. Now, the anticipated latency improvement can be seen. Compared to 159cefb14d9f951ae79ec8265c6bb0395009b19c, the latency for samples with 16 values or less is up to 33% better.

Furthermore, the check for signaling is fixed and some typos are fixed.