unit-tests: intialize global test params at runtime
Move the intialization of the paramters to runtime by converting the
static array to a vector and filling it when critirion calls the test
setup.

Global variable initialization order across translation units is not
defined.
Referring to e.g. the memory::mmap global variable in the initialization
of the static unit test parameter set may leave an incorrect address in
the struct param.

Signed-off-by: Philipp Jungkamp <Philipp.Jungkamp@opal-rt.com>