Added Tests for the Python-Wrapper
The tests are not yet integrated into the pipeline and are primarily focused on the signal_v2 node.
Not all functions work as expected.

node_name_short() appears to be print a freed string.

node_stop() and node_restart() do not function properly.
However, they both use the same underlying function to stop a node, which may
be the cause of their failure.

Signed-off-by: Kevin Vu te Laar <vu.te@rwth-aachen.de>