unit tests: add test and enable test discovery
Unit Tests in `tests/unit/python/` are now executed automatically as long as
their filenames start with the prefix `test`.

The binding unit test was renamed/moved from
`tests/unit/python_binding.py` to
`tests/unit/python/test_python_binding.py`.

Node implementations return `-1` for functions that are not implemented.
This is now covered for the `node_reverse()` function of the `socket`
node.

Signed-off-by: Kevin Vu te Laar <vu.te@rwth-aachen.de>